+++ src/test/java/security/SecurityTestTask.kt
@@ -2,6 +2,7 @@ package security
 
 import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.LocalQuickFix
+import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.psi.PsiElement
@@ -21,6 +22,7 @@ open class SecurityTestTask: BasePlatformTestCase() {
         ApplicationManager.getApplication().runReadAction {
             val mockHolder = mock<ProblemsHolder> {
                 on { registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>()) } doAnswer {}
+                on { registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>(), anyVararg<LocalQuickFix>()) } doAnswer {}
             }
             val testFile = this.createLightFile(filename, PythonFileType.INSTANCE.language, code);
             val mockLocalSession = mock<LocalInspectionToolSession> {
@@ -34,7 +36,11 @@ open class SecurityTestTask: BasePlatformTestCase() {
             expr.forEach { e ->
                 testVisitor.visitPyAssignmentStatement(e)
             }
-            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>())
+            try {
+                Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>())
+            } catch (a: AssertionError){
+                Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>(), anyVararg<LocalQuickFix>())
+            }
             Mockito.verify(mockLocalSession, Mockito.times(1)).file
         }
     }
@@ -43,6 +49,7 @@ open class SecurityTestTask: BasePlatformTestCase() {
         ApplicationManager.getApplication().runReadAction {
             val mockHolder = mock<ProblemsHolder> {
                 on { registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>()) } doAnswer {}
+                on { registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>(), anyVararg<LocalQuickFix>()) } doAnswer {}
             }
             val testFile = this.createLightFile(filename, PythonFileType.INSTANCE.language, code);
             val mockLocalSession = mock<LocalInspectionToolSession> {
@@ -56,7 +63,11 @@ open class SecurityTestTask: BasePlatformTestCase() {
             expr.forEach { e ->
                 testVisitor.visitPyCallExpression(e)
             }
-            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>())
+            try {
+                Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>())
+            } catch (a: AssertionError){
+                Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>(), anyVararg<LocalQuickFix>())
+            }
             Mockito.verify(mockLocalSession, Mockito.times(1)).file
         }
     }
@@ -123,6 +134,50 @@ open class SecurityTestTask: BasePlatformTestCase() {
                 testVisitor.visitPyFormattedStringElement(e)
             }
             Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), anyVararg<LocalQuickFix>())
+            Mockito.verify(mockLocalSession, Mockito.times(1)).file
+        }
+    }
+
+    inline fun <inspector: PyInspection>testAssertStatement(code: String, times: Int = 1, check: Checks.CheckType, filename: String = "test.py", instance: inspector){
+        ApplicationManager.getApplication().runReadAction {
+            val mockHolder = mock<ProblemsHolder> {
+                on { registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>()) } doAnswer {}
+            }
+            val testFile = this.createLightFile(filename, PythonFileType.INSTANCE.language, code);
+            val mockLocalSession = mock<LocalInspectionToolSession> {
+                on { file } doReturn (testFile)
+            }
+            assertNotNull(testFile)
+            val testVisitor = instance.buildVisitor(mockHolder, true, mockLocalSession) as PyInspectionVisitor
+
+            val expr: @NotNull MutableCollection<PyAssertStatement> = PsiTreeUtil.findChildrenOfType(testFile, PyAssertStatement::class.java)
+            assertNotNull(expr)
+            expr.forEach { e ->
+                testVisitor.visitPyAssertStatement(e)
+            }
+            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>())
+            Mockito.verify(mockLocalSession, Mockito.times(1)).file
+        }
+    }
+
+    inline fun <inspector: PyInspection>testTryExceptStatement(code: String, times: Int = 1, check: Checks.CheckType, filename: String = "test.py", instance: inspector){
+        ApplicationManager.getApplication().runReadAction {
+            val mockHolder = mock<ProblemsHolder> {
+                on { registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>()) } doAnswer {}
+            }
+            val testFile = this.createLightFile(filename, PythonFileType.INSTANCE.language, code);
+            val mockLocalSession = mock<LocalInspectionToolSession> {
+                on { file } doReturn (testFile)
+            }
+            assertNotNull(testFile)
+            val testVisitor = instance.buildVisitor(mockHolder, true, mockLocalSession) as PyInspectionVisitor
+
+            val expr: @NotNull MutableCollection<PyTryExceptStatement> = PsiTreeUtil.findChildrenOfType(testFile, PyTryExceptStatement::class.java)
+            assertNotNull(expr)
+            expr.forEach { e ->
+                testVisitor.visitPyTryExceptStatement(e)
+            }
+            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(check.getDescription()), any<ProblemHighlightType>())
             Mockito.verify(mockLocalSession, Mockito.times(1)).file
         }
     }